test(upload planner): Move stub upload stats to shared stubs PE-460
